Implement PackageSearchResults component to handle loading, error, empty, and package display states in trending page. Refactor trending.tsx to utilize the new component, improving code organization and readability.
